# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BitBox API",
    "description": "Rust API for 360 Project notation online",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/admin/config": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get main app configuration",
        "description": "This endpoint returns the main configuration of Bitbox.",
        "operationId": "get_main_config_route",
        "responses": {
          "200": {
            "description": "The main config of the application",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainConfig"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "examples": {
                  "InvalidToken": {
                    "value": "Error"
                  },
                  "NoToken": {
                    "value": "Token not provided"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "Forbidden"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update the main configuration",
        "description": "This endpoint updates the main configuration of the application.",
        "operationId": "update_main_config_route",
        "requestBody": {
          "description": "The updated main configuration object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatedMainConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Main configuration updated successfully"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                },
                "examples": {
                  "InvalidToken": {
                    "value": "Error"
                  },
                  "NoToken": {
                    "value": "Token not provided"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "Forbidden"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                },
                "example": "Database record"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalError"
                },
                "example": "InternalError"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InternalError": {
        "default": null,
        "nullable": true
      },
...
      "MainConfig": {
        "type": "object",
        "required": [
          "id",
          "register",
          "authorized_domains",
          "updated_at"
        ],
        "properties": {
          "authorized_domains": {
            "type": "array",
            "items": {
              "type": "string",
              "nullable": true
            }
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "register": {
            "type": "boolean"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "NotFoundError": {
        "type": "object",
        "required": [
          "resource"
        ],
        "properties": {
          "resource": {
            "type": "string"
          }
        }
      },
...
      "UnauthorizedError": {
        "default": null,
        "nullable": true
      },
...
      "UpdatedMainConfig": {
        "type": "object",
        "properties": {
          "authorized_domains": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "register": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "api_key": [
        "read",
        "write"
      ]
    }
  ]
}